+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project will adhere to [Semantic Versioning](http://semver.org/spec/v2.
 
 ## [Unreleased]
 
+### Fixed
+
+* Always return predicates of list type in an array.
+* Edges without facet values are also returned when performing sort on facet.
+* Dont derive schema while deleting edges.
+
+### Changed
+
+* Remove StartTs field from `api.Operation`.
+
 ## [1.0.1] - 2017-12-20
 
 ### Fixed